Update cato_sase.md
My name is Yaakov and I’m the Doc Manager for Cato Networks. We recently updated the CMA paths and I made changes to reflect the new navigation. (the screenshots are still out of date)
 
The correct information is in this Cato article: https://support.catonetworks.com/hc/en-us/articles/4413280536081-Generating-API-Keys-for-the-Cato-API
 
Thanks! yaakov.simon@catonetworks.com